requirements.txt
torch>=2.2
numpy>=1.24
pandas>=2.0
pyyaml>=6.0
scikit-learn>=1.3
matplotlib>=3.8
pillow>=10.0
streamlit>=1.31
fastapi>=0.110
uvicorn>=0.27
onnx>=1.16
onnxruntime>=1.17
tqdm>=4.66
